% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geokmeans-package.R
\docType{package}
\name{geokmeans-package}
\alias{geokmeans}
\alias{geokmeans-package}
\title{geokmeans: Fast and Eco-Friendly k-Means Clustering Algorithms}
\description{
Fast C++ implementations of several k-means clustering algorithms exposed to
R through a uniform interface: Lloyd's algorithm, Elkan, Hamerly, Annulus,
Exponion, Ball k-means, and the bound-free Geometric-k-means method.
}
\details{
The main entry points are \code{\link[=geo_kmeans]{geo_kmeans()}}, \code{\link[=lloyd_kmeans]{lloyd_kmeans()}}, \code{\link[=elkan_kmeans]{elkan_kmeans()}},
\code{\link[=hamerly_kmeans]{hamerly_kmeans()}}, \code{\link[=annulus_kmeans]{annulus_kmeans()}}, \code{\link[=exponion_kmeans]{exponion_kmeans()}}, \code{\link[=ball_kmeans]{ball_kmeans()}},
and the dispatcher \code{\link[=kmeans_dc]{kmeans_dc()}}.
}
\references{
Sharma, P., Stanislaw, M., Kurban, H., Kulekci, O., and Dalkilic, M. (2025).
Geometric-k-means: A Bound Free Approach to Fast and Eco-Friendly k-means.
\doi{10.1007/s10994-025-06891-1}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/parichit/Geometric-k-means}
  \item Report bugs at \url{https://github.com/parichit/Geometric-k-means/issues}
}

}
\author{
\strong{Maintainer}: Parichit Sharma \email{parishar@iu.edu} [copyright holder]

Authors:
\itemize{
  \item Marcin Stanislaw
  \item Hasan Kurban
  \item Oguzhan Kulekci
  \item Mehmet Dalkilic
}

}
\keyword{internal}
